Refactor all patch functions into separate package
